backend/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
blis==0.7.11
boto3
botocore==1.34.85
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.15
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
coloredlogs==15.0.1
confection==0.1.4
cssselect==1.2.0
cymem==2.0.8
dataclasses-json==0.6.4
Deprecated==1.2.14
deprecation==2.1.0
distro==1.9.0
dnspython==2.6.1
docker==7.0.0
duckduckgo-search==3.5.0
email_validator==2.1.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
fastapi==0.110.1
filelock==3.13.4
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.31
google-api-core==2.18.0
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
gotrue==2.4.2
grpcio==1.62.1
gTTS==2.3.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.7
importlib-metadata==7.0.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.3
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kubernetes==29.0.0
langchain==0.1.16
langchain-community==0.0.33
langchain-core==0.1.43
langchain-openai==0.1.3
langchain-text-splitters==0.0.1
langcodes==3.3.0
langsmith==0.1.48
loguru==0.7.2
lxml==5.3.0
MarkupSafe==2.1.5
marshmallow==3.21.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.6.0
# numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.3
openai==1.20.0
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
orjson==3.10.1
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pillow==10.3.0
pinecone-client==2.2.1
playsound==1.2.2
postgrest==0.16.4
posthog==3.5.0
preshed==3.0.9
proto-plus==1.23.0
protobuf==4.25.3
psycopg2==2.9.9
pulsar-client==3.5.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.0
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.18.1
pyparsing==3.1.2
PyPika==0.48.9
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.9
PyYAML==6.0.1
readability-lxml==0.8.1
realtime==1.0.4
redis==5.0.3
referencing==0.34.0
regex==2024.4.16
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.18.0
rsa==4.9
s3transfer==0.10.1
selenium==4.19.0
semanticscholar==0.8.4
setuptools==69.5.1
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.29
srsly==2.4.8
starlette==0.37.2
storage3==0.7.4
StrEnum==0.4.15
supabase==2.4.5
supafunc==0.4.5
sympy==1.12
tenacity==8.2.3
thinc==8.2.3
tiktoken==0.7.0
tokenizers==0.15.2
tqdm==4.66.2
trio==0.25.0
trio-websocket==0.11.1
tweepy==4.14.0
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.11.0
ujson==5.9.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
wasabi==1.1.2
watchfiles==0.21.0
weasel==0.3.4
webdriver-manager==4.0.1
websocket-client==1.7.0
websockets==12.0
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.18.1

json_repair
text_generation
transformers
faiss-cpu==1.8.0
pyautogen
Collapse
xinference-client==0.14.0
cohere
pyjwt==2.9.0
redis
cryptography==43.0.1
celery==5.3.6
flower==2.0.1
google-auth-oauthlib 
oauth2client
moviepy
pymupdf4llm